Power pellet updates
If we are scared, put both to offense. 
If they are scared, our pacman will go into a deadend for a pellet, if there is time to get out of the deadend before the closest ghost becomes not scared. 
If they are scared, and they are in our way, our pacman will run over them, but won't chase them (this one might need some testing, I don't know why sometimes the pacman runs away still).